Moved MessageFormatter::Builder implementation to separate file; changed Operand representation
