+++ icu4c/source/i18n/messageformat2.cpp
@@ -23,7 +23,6 @@ using Operand         = MessageFormatDataModel::Operand;
 using Operator        = MessageFormatDataModel::Operator;
 using Pattern         = MessageFormatDataModel::Pattern;
 using PatternPart     = MessageFormatDataModel::PatternPart;
-using Reserved        = MessageFormatDataModel::Reserved;
 using SelectorKeys    = MessageFormatDataModel::SelectorKeys;
 using VariantMap      = MessageFormatDataModel::VariantMap;
 
@@ -31,313 +30,19 @@ using PrioritizedVariantList = ImmutableVector<PrioritizedVariant>;
 
 #define TEXT_SELECTOR UnicodeString("select")
 
-// -------------------------------------
-// Creates a MessageFormat instance based on the pattern.
-
-// Returns a new (uninitialized) builder
-MessageFormatter::Builder* MessageFormatter::builder(UErrorCode& errorCode) {
-    if (U_FAILURE(errorCode)) {
-        return nullptr;
-    }
-    LocalPointer<MessageFormatter::Builder> tree(new Builder());
-    if (!tree.isValid()) {
-        errorCode = U_MEMORY_ALLOCATION_ERROR;
-        return nullptr;
-    }
-    return tree.orphan();
-}
-
-MessageFormatter::Builder& MessageFormatter::Builder::setPattern(const UnicodeString& pat, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    pattern.adoptInstead(new UnicodeString(pat));
-    if (!pattern.isValid()) {
-        errorCode = U_MEMORY_ALLOCATION_ERROR;
-    } else {
-        // Invalidate the data model
-        dataModel.adoptInstead(nullptr);
-    }
-    return *this;
-}
-
-// Precondition: `reg` is non-null
-// Does *not* adopt `reg`
-MessageFormatter::Builder& MessageFormatter::Builder::setFunctionRegistry(const FunctionRegistry* reg) {
-    U_ASSERT(reg != nullptr);
-    customFunctionRegistry = reg;
-    return *this;
-}
-
-MessageFormatter::Builder& MessageFormatter::Builder::setLocale(const Locale& loc) {
-    locale = loc;
-    return *this;
-}
-
-// Takes ownership of `dataModel`
-MessageFormatter::Builder& MessageFormatter::Builder::setDataModel(MessageFormatDataModel* newDataModel) {
-  U_ASSERT(newDataModel != nullptr);
-  dataModel.adoptInstead(newDataModel);
-
-  // Invalidate the pattern
-  pattern.adoptInstead(nullptr);
-  return *this;
-}
-
-/*
-  For now, this is a destructive build(); it invalidates the builder
-
-  It's probably better for this method to either copy the builder,
-  or use a reference to (e.g.) a function registry with the assumption that it
-  has the same lifetime as the formatter.
-  See the custom functions example in messageformat2test.cpp for motivation.
-*/
-MessageFormatter* MessageFormatter::Builder::build(UParseError& parseError, UErrorCode& errorCode) {
-    NULL_ON_ERROR(errorCode);
-
-    LocalPointer<MessageFormatter> mf(new MessageFormatter(*this, parseError, errorCode));
-    if (U_FAILURE(errorCode)) {
-        return nullptr;
-    }
-    return mf.orphan();
-}
-
-void MessageFormatter::initErrors(UErrorCode& errorCode) {
-    CHECK_ERROR(errorCode);
-    errors.adoptInstead(Errors::create(errorCode));
-}
-
-MessageFormatter::MessageFormatter(MessageFormatter::Builder& builder, UParseError &parseError,
-                                   UErrorCode &success) : locale(builder.locale), customFunctionRegistry(builder.customFunctionRegistry) {
-    CHECK_ERROR(success);
-
-    // Set up the standard function registry
-    LocalPointer<FunctionRegistry::Builder> standardFunctionsBuilder(FunctionRegistry::builder(success));
-    CHECK_ERROR(success);
-
-    standardFunctionsBuilder->setFormatter(UnicodeString("datetime"), new StandardFunctions::DateTimeFactory(), success)
-        .setFormatter(UnicodeString("number"), new StandardFunctions::NumberFactory(), success)
-        .setFormatter(UnicodeString("identity"), new StandardFunctions::IdentityFactory(), success)
-        .setSelector(UnicodeString("plural"), new StandardFunctions::PluralFactory(UPLURAL_TYPE_CARDINAL), success)
-        .setSelector(UnicodeString("selectordinal"), new StandardFunctions::PluralFactory(UPLURAL_TYPE_ORDINAL), success)
-        .setSelector(UnicodeString("select"), new StandardFunctions::TextFactory(), success)
-        .setSelector(UnicodeString("gender"), new StandardFunctions::TextFactory(), success);
-    standardFunctionRegistry.adoptInstead(standardFunctionsBuilder->build(success));
-    CHECK_ERROR(success);
-    standardFunctionRegistry->checkStandard();
-
-    initErrors(success);
-    CHECK_ERROR(success);
-
-    // Validate pattern and build data model
-    // First, check that exactly one of the pattern and data model are set, but not both
-
-    bool patternSet = builder.pattern.isValid();
-    bool dataModelSet = builder.dataModel.isValid();
-
-    if ((!patternSet && !dataModelSet)
-        || (patternSet && dataModelSet)) {
-      success = U_INVALID_STATE_ERROR;
-      return;
-    }
-
-    // If data model was set, just assign it
-    if (dataModelSet) {
-      dataModel.adoptInstead(builder.dataModel.orphan());
-      return;
-    }
-
-    LocalPointer<MessageFormatDataModel::Builder> tree(MessageFormatDataModel::builder(success));
-    if (U_FAILURE(success)) {
-      return;
-    }
-
-    // Initialize formatter cache
-    cachedFormatters.adoptInstead(new CachedFormatters(success));
-
-    // Parse the pattern
-    normalizedInput.adoptInstead(new UnicodeString(u""));
-    if (!normalizedInput.isValid()) {
-      success = U_MEMORY_ALLOCATION_ERROR;
-      return;
-    }
-    LocalPointer<Parser> parser(Parser::create(*builder.pattern, *tree, *normalizedInput, *errors, success));
-    CHECK_ERROR(success);
-    parser->parse(parseError, success);
-
-    // Build the data model based on what was parsed
-    LocalPointer<MessageFormatDataModel> dataModelPtr(tree->build(success));
-    if (U_SUCCESS(success)) {
-      dataModel.adoptInstead(dataModelPtr.orphan());
-    }
-}
-
-MessageFormatDataModel::~MessageFormatDataModel() {}
-MessageFormatter::~MessageFormatter() {}
-
-// ------------------------------------------------------
-// MessageArguments
-
-using Arguments = MessageArguments;
-
-bool Arguments::has(const VariableName& arg) const {
-    U_ASSERT(contents.isValid() && objectContents.isValid());
-    return contents->containsKey(arg.name()) || objectContents->containsKey(arg.name());
-}
-
-const Formattable& Arguments::get(const VariableName& arg) const {
-    U_ASSERT(has(arg));
-    const Formattable* result = static_cast<const Formattable*>(contents->get(arg.name()));
-    if (result == nullptr) {
-        result = static_cast<const Formattable*>(objectContents->get(arg.name()));
-    }
-    U_ASSERT(result != nullptr);
-    return *result;
-}
-
-Arguments::Builder::Builder(UErrorCode& errorCode) {
-    CHECK_ERROR(errorCode);
-
-    contents.adoptInstead(new Hashtable(compareVariableName, nullptr, errorCode));
-    objectContents.adoptInstead(new Hashtable(compareVariableName, nullptr, errorCode));
-    CHECK_ERROR(errorCode);
-    // The `contents` hashtable owns the values, but does not own the keys
-    contents->setValueDeleter(uprv_deleteUObject);
-    // The `objectContents` hashtable does not own the values
-}
-
-Arguments::Builder& Arguments::Builder::add(const UnicodeString& name, const UnicodeString& val, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    Formattable* valPtr(ExpressionContext::createFormattable(val, errorCode));
-    THIS_ON_ERROR(errorCode);
-    return add(name, valPtr, errorCode);
-}
-
-Arguments::Builder& Arguments::Builder::addDouble(const UnicodeString& name, double val, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    Formattable* valPtr(ExpressionContext::createFormattable(val, errorCode));
-    THIS_ON_ERROR(errorCode);
-    return add(name, valPtr, errorCode);
-}
-
-Arguments::Builder& Arguments::Builder::addInt64(const UnicodeString& name, int64_t val, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    Formattable* valPtr(ExpressionContext::createFormattable(val, errorCode));
-    THIS_ON_ERROR(errorCode);
-    return add(name, valPtr, errorCode);
-}
-
-Arguments::Builder& Arguments::Builder::addDate(const UnicodeString& name, UDate val, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    Formattable* valPtr(ExpressionContext::createFormattableDate(val, errorCode));
-    THIS_ON_ERROR(errorCode);
-    return add(name, valPtr, errorCode);
-}
-
-Arguments::Builder& Arguments::Builder::addDecimal(const UnicodeString& name, StringPiece val, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    Formattable* valPtr(ExpressionContext::createFormattableDecimal(val, errorCode));
-    THIS_ON_ERROR(errorCode);
-    return add(name, valPtr, errorCode);
-}
-
-Arguments::Builder& Arguments::Builder::add(const UnicodeString& name, const UnicodeString* arr, int32_t count, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    Formattable* valPtr(ExpressionContext::createFormattable(arr, count, errorCode));
-    THIS_ON_ERROR(errorCode);
-    return add(name, valPtr, errorCode);
-}
-
-// Does not adopt the object
-Arguments::Builder& Arguments::Builder::addObject(const UnicodeString& name, const UObject* obj, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    // The const_cast is valid because the object will only be accessed via
-    // getObjectInput(), which returns a const reference
-    Formattable* valPtr(ExpressionContext::createFormattable(const_cast<UObject*>(obj), errorCode));
-    THIS_ON_ERROR(errorCode);
-    objectContents->put(name, valPtr, errorCode);
-    return *this;
-}
-
-// Adopts its argument
-Arguments::Builder& Arguments::Builder::add(const UnicodeString& name, Formattable* value, UErrorCode& errorCode) {
-    THIS_ON_ERROR(errorCode);
-
-    U_ASSERT(value != nullptr);
-
-    contents->put(name, value, errorCode);
-    return *this;
-}
-
-/* static */ MessageArguments::Builder* MessageArguments::builder(UErrorCode& errorCode) {
-    NULL_ON_ERROR(errorCode);
-    MessageArguments::Builder* result = new MessageArguments::Builder(errorCode);
-    if (result == nullptr) {
-        errorCode = U_MEMORY_ALLOCATION_ERROR;
-    }
-    return result;
-}
-
-MessageArguments* MessageArguments::Builder::build(UErrorCode& errorCode) const {
-    NULL_ON_ERROR(errorCode);
-    U_ASSERT(contents.isValid() && objectContents.isValid());
-
-    LocalPointer<Hashtable> contentsCopied(new Hashtable(compareVariableName, nullptr, errorCode));
-    LocalPointer<Hashtable> objectContentsCopied(new Hashtable(compareVariableName, nullptr, errorCode));
-    NULL_ON_ERROR(errorCode);
-    // The `contents` hashtable owns the values, but does not own the keys
-    contents->setValueDeleter(uprv_deleteUObject);
-    // The `objectContents` hashtable does not own the values
-
-    int32_t pos = UHASH_FIRST;
-    LocalPointer<Formattable> optionValue;
-    // Copy the non-objects
-    while (true) {
-        const UHashElement* element = contents->nextElement(pos);
-        if (element == nullptr) {
-            break;
-        }
-        const Formattable& toCopy = *(static_cast<Formattable*>(element->value.pointer));
-        optionValue.adoptInstead(new Formattable(toCopy));
-        if (!optionValue.isValid()) {
-            errorCode = U_MEMORY_ALLOCATION_ERROR;
-            return nullptr;
-        }
-        UnicodeString* key = static_cast<UnicodeString*>(element->key.pointer);
-        contentsCopied->put(*key, optionValue.orphan(), errorCode);
-    }
-    // Copy the objects
-    pos = UHASH_FIRST;
-    while (true) {
-        const UHashElement* element = objectContents->nextElement(pos);
-        if (element == nullptr) {
-            break;
-        }
-        UnicodeString* key = static_cast<UnicodeString*>(element->key.pointer);
-        objectContentsCopied->put(*key, element->value.pointer, errorCode);
-    }
-    MessageArguments* result = new MessageArguments(contentsCopied.orphan(), objectContentsCopied.orphan());
-    if (result == nullptr) {
-        errorCode = U_MEMORY_ALLOCATION_ERROR;
-    }
-    return result;
-}
-
 // ------------------------------------------------------
 // Formatting
 
+// The result of formatting a literal is just itself.
 static const Formattable& evalLiteral(const Literal& lit) {
     return lit.getContents();
 }
 
+// Assumes that `var` is a message argument; sets the input in the context
+// to the argument's value.
 void MessageFormatter::evalArgument(const VariableName& var, ExpressionContext& context) const {
     U_ASSERT(context.hasGlobal(var));
+    // The fallback for a variable name is itself.
     context.setFallback(var);
     if (context.hasGlobalAsFormattable(var)) {
         context.setInput(context.getGlobalAsFormattable(var));
@@ -346,7 +51,9 @@ void MessageFormatter::evalArgument(const VariableName& var, ExpressionContext& 
     }
 }
 
+// Sets the input to the contents of the literal
 void MessageFormatter::formatLiteral(const Literal& lit, ExpressionContext& context) const {
+    // The fallback for a literal is itself.
     context.setFallback(lit);
     context.setInput(evalLiteral(lit));
 }
@@ -363,7 +70,7 @@ void MessageFormatter::formatOperand(const Environment& env, const Operand& rand
         // resolution of:
         //   https://github.com/unicode-org/message-format-wg/issues/310
         // it might need to forbid it.
-        const VariableName& var = *rand.asVariable();
+        const VariableName& var = rand.asVariable();
         // TODO: Currently, this code implements lazy evaluation of locals.
         // That is, the environment binds names to an Expression, not a resolved value.
         // Eager vs. lazy evaluation is an open issue:
@@ -386,7 +93,7 @@ void MessageFormatter::formatOperand(const Environment& env, const Operand& rand
             return;
         }
     } else if (rand.isLiteral()) {
-        formatLiteral(*rand.asLiteral(), context);
+        formatLiteral(rand.asLiteral(), context);
         return;
     }
 }
@@ -502,10 +209,10 @@ void MessageFormatter::resolveVariables(const Environment& env, const Operand& r
         return;
     } else if (rand.isLiteral()) {
         U_ASSERT(!context.hasOperator());
-        formatLiteral(*rand.asLiteral(), context);
+        formatLiteral(rand.asLiteral(), context);
     } else {
         // Must be variable
-        const VariableName& var = *rand.asVariable();
+        const VariableName& var = rand.asVariable();
         // Resolve it
         const Closure* referent = env.lookup(var);
         if (referent != nullptr) {
@@ -981,7 +688,7 @@ void MessageFormatter::check(Context& context, const Environment& localEnv, cons
     }
 
     // Check that variable is in scope
-    const VariableName& var = *rand.asVariable();
+    const VariableName& var = rand.asVariable();
     // Check local scope
     if (localEnv.lookup(var) != nullptr) {
         return;
@@ -1029,7 +736,7 @@ void MessageFormatter::checkDeclarations(Context& context, Environment*& env, UE
     }
 }
 
-void MessageFormatter::formatToString(const Arguments& arguments, UErrorCode &status, UnicodeString& result) const {
+void MessageFormatter::formatToString(const MessageArguments& arguments, UErrorCode &status, UnicodeString& result) const {
     CHECK_ERROR(status);
 
     // Note: we currently evaluate variables lazily,
