+++ icu4c/source/i18n/messageformat2_builder.cpp
@@ -1,0 +1,181 @@
+// © 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
+
+#include "unicode/utypes.h"
+
+#if !UCONFIG_NO_FORMATTING
+
+#include "unicode/messageformat2.h"
+#include "unicode/messageformat2_data_model.h"
+#include "uvector.h" // U_ASSERT
+
+U_NAMESPACE_BEGIN namespace message2 {
+
+/*
+using Binding         = MessageFormatDataModel::Binding;
+using Bindings        = MessageFormatDataModel::Bindings;
+using Expression      = MessageFormatDataModel::Expression;
+using ExpressionList  = MessageFormatDataModel::ExpressionList;
+using Key             = MessageFormatDataModel::Key;
+using KeyList         = MessageFormatDataModel::KeyList;
+using Literal         = MessageFormatDataModel::Literal;
+using OptionMap       = MessageFormatDataModel::OptionMap;
+using Operand         = MessageFormatDataModel::Operand;
+using Operator        = MessageFormatDataModel::Operator;
+using Pattern         = MessageFormatDataModel::Pattern;
+using PatternPart     = MessageFormatDataModel::PatternPart;
+using SelectorKeys    = MessageFormatDataModel::SelectorKeys;
+using VariantMap      = MessageFormatDataModel::VariantMap;
+
+using PrioritizedVariantList = ImmutableVector<PrioritizedVariant>;
+
+#define TEXT_SELECTOR UnicodeString("select")
+*/
+
+// -------------------------------------
+// Creates a MessageFormat instance based on the pattern.
+
+// Returns a new (uninitialized) builder
+MessageFormatter::Builder* MessageFormatter::builder(UErrorCode& errorCode) {
+    if (U_FAILURE(errorCode)) {
+        return nullptr;
+    }
+    LocalPointer<MessageFormatter::Builder> tree(new Builder());
+    if (!tree.isValid()) {
+        errorCode = U_MEMORY_ALLOCATION_ERROR;
+        return nullptr;
+    }
+    return tree.orphan();
+}
+
+MessageFormatter::Builder& MessageFormatter::Builder::setPattern(const UnicodeString& pat, UErrorCode& errorCode) {
+    THIS_ON_ERROR(errorCode);
+
+    pattern.adoptInstead(new UnicodeString(pat));
+    if (!pattern.isValid()) {
+        errorCode = U_MEMORY_ALLOCATION_ERROR;
+    } else {
+        // Invalidate the data model
+        dataModel.adoptInstead(nullptr);
+    }
+    return *this;
+}
+
+// Precondition: `reg` is non-null
+// Does *not* adopt `reg`
+MessageFormatter::Builder& MessageFormatter::Builder::setFunctionRegistry(const FunctionRegistry* reg) {
+    U_ASSERT(reg != nullptr);
+    customFunctionRegistry = reg;
+    return *this;
+}
+
+MessageFormatter::Builder& MessageFormatter::Builder::setLocale(const Locale& loc) {
+    locale = loc;
+    return *this;
+}
+
+// Takes ownership of `dataModel`
+MessageFormatter::Builder& MessageFormatter::Builder::setDataModel(MessageFormatDataModel* newDataModel) {
+  U_ASSERT(newDataModel != nullptr);
+  dataModel.adoptInstead(newDataModel);
+
+  // Invalidate the pattern
+  pattern.adoptInstead(nullptr);
+  return *this;
+}
+
+/*
+  For now, this is a destructive build(); it invalidates the builder
+
+  It's probably better for this method to either copy the builder,
+  or use a reference to (e.g.) a function registry with the assumption that it
+  has the same lifetime as the formatter.
+  See the custom functions example in messageformat2test.cpp for motivation.
+*/
+MessageFormatter* MessageFormatter::Builder::build(UParseError& parseError, UErrorCode& errorCode) {
+    NULL_ON_ERROR(errorCode);
+
+    LocalPointer<MessageFormatter> mf(new MessageFormatter(*this, parseError, errorCode));
+    if (U_FAILURE(errorCode)) {
+        return nullptr;
+    }
+    return mf.orphan();
+}
+
+void MessageFormatter::initErrors(UErrorCode& errorCode) {
+    CHECK_ERROR(errorCode);
+    errors.adoptInstead(Errors::create(errorCode));
+}
+
+MessageFormatter::MessageFormatter(MessageFormatter::Builder& builder, UParseError &parseError,
+                                   UErrorCode &success) : locale(builder.locale), customFunctionRegistry(builder.customFunctionRegistry) {
+    CHECK_ERROR(success);
+
+    // Set up the standard function registry
+    LocalPointer<FunctionRegistry::Builder> standardFunctionsBuilder(FunctionRegistry::builder(success));
+    CHECK_ERROR(success);
+
+    standardFunctionsBuilder->setFormatter(UnicodeString("datetime"), new StandardFunctions::DateTimeFactory(), success)
+        .setFormatter(UnicodeString("number"), new StandardFunctions::NumberFactory(), success)
+        .setFormatter(UnicodeString("identity"), new StandardFunctions::IdentityFactory(), success)
+        .setSelector(UnicodeString("plural"), new StandardFunctions::PluralFactory(UPLURAL_TYPE_CARDINAL), success)
+        .setSelector(UnicodeString("selectordinal"), new StandardFunctions::PluralFactory(UPLURAL_TYPE_ORDINAL), success)
+        .setSelector(UnicodeString("select"), new StandardFunctions::TextFactory(), success)
+        .setSelector(UnicodeString("gender"), new StandardFunctions::TextFactory(), success);
+    standardFunctionRegistry.adoptInstead(standardFunctionsBuilder->build(success));
+    CHECK_ERROR(success);
+    standardFunctionRegistry->checkStandard();
+
+    initErrors(success);
+    CHECK_ERROR(success);
+
+    // Validate pattern and build data model
+    // First, check that exactly one of the pattern and data model are set, but not both
+
+    bool patternSet = builder.pattern.isValid();
+    bool dataModelSet = builder.dataModel.isValid();
+
+    if ((!patternSet && !dataModelSet)
+        || (patternSet && dataModelSet)) {
+      success = U_INVALID_STATE_ERROR;
+      return;
+    }
+
+    // If data model was set, just assign it
+    if (dataModelSet) {
+      dataModel.adoptInstead(builder.dataModel.orphan());
+      return;
+    }
+
+    LocalPointer<MessageFormatDataModel::Builder> tree(MessageFormatDataModel::builder(success));
+    if (U_FAILURE(success)) {
+      return;
+    }
+
+    // Initialize formatter cache
+    cachedFormatters.adoptInstead(new CachedFormatters(success));
+
+    // Parse the pattern
+    normalizedInput.adoptInstead(new UnicodeString(u""));
+    if (!normalizedInput.isValid()) {
+      success = U_MEMORY_ALLOCATION_ERROR;
+      return;
+    }
+    LocalPointer<Parser> parser(Parser::create(*builder.pattern, *tree, *normalizedInput, *errors, success));
+    CHECK_ERROR(success);
+    parser->parse(parseError, success);
+
+    // Build the data model based on what was parsed
+    LocalPointer<MessageFormatDataModel> dataModelPtr(tree->build(success));
+    if (U_SUCCESS(success)) {
+      dataModel.adoptInstead(dataModelPtr.orphan());
+    }
+}
+
+MessageFormatDataModel::~MessageFormatDataModel() {}
+MessageFormatter::~MessageFormatter() {}
+
+} // namespace message2
+U_NAMESPACE_END
+
+#endif /* #if !UCONFIG_NO_FORMATTING */
